requirements.txt
# Streamlit
streamlit==1.40.1

# LangChain (Groq + HuggingFace Stack)
langchain
langchain-core
langchain-community==0.3.5
langchain-groq
langchain-chroma
langchain-huggingface

# VectorDB
chromadb==0.5.20

# Data & Config
pydantic==2.10.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
python-dateutil==2.9.0